www/run_checks: remove useless prefix
A prefix "zselenium-" was added to the hostname when launching the web
tests. This was originally intended to avoid interacting with the httpd
configuration already present on developers' machines, but:
* it results in uselessly long serverName used for the test
(geosafe-ea-cm.zselenium.geosafe-ea.builbot.build is quite long)
* as the same variable is used for ssl certificates path, keeping it
would also have implied having uselessly long paths for theses
certificates
* it does not really conflict with the devs' httpd conf file anyway
(and hardly anyone launches the web tests using waf on their dev
machine)

So we remove this prefix.

Change-Id: I8941a360c53c75be1ae6af38fa70d83f4c3743fa
Priv-Id: abda9b71c9589d1f04ae331be0814f1a1b7afda4