qpsstress: fuzzing on snapshots and reopen of QPS
Even if it breaks the philosophy of fuzzing in terms of executions per
second and efficiency, implement snapshot and reopen in fuzzing
steps, so we can have a better test coverage of the whole QPS thing.

Improvements on fuzzing when opening/compressing/writing operations are
performed on QPS database are mandatory to consider a good fuzzing
approach here.

Change-Id: If925f1ba1b104597e8caae09534f41fa4a1754d7
Priv-Id: 8dd7c5a7662f06ff05bf1dccc7b4d87f0b9460a1