qpsstress: split fuzzing operations
Implementation of fuzzing operations before this patch is not efficient
enough because:
 - Snapshot and reopen operations kill the purpose of fuzzing (slow
   operations)
 - QPS objects and all related operations are based on QPS allocator but
   don't really interact with QPS code except for creating/accesing
   handles and QPS pages, accessing/clearing data inside QPS.

Splitting fuzzing operations into different subgroups forces the fuzzer
to focus on each category and tries then to perform better combinations
of operations inside these categories. It can now be performed with the
following environment variable:
    FUZZING_CAT=value

Current available categories are:
    0: no filter on categories (this is the default choice if FUZZING_CAT
       is not available
    1: focus fuzzing on QPS allocator
    2: focus fuzzing on QPS objects and associated operations

Even if there is still a chance that QPS object category could interfere
with the QPS allocator one, the fuzzer should still be able to recreate
issues, by performing more operations to achieve the same result.

Change-Id: I5a48895a29bfe848df120df3b6b56e3e84d07d50
Priv-Id: 6b665ab8eaec4da2f4f63474cf225a97d38f030a