iopc: Remove --typescript-enable-backbone option (always enable it).
This option was added because when it was introduced, iopc was in the
tools and we had to maintain backward compatibility.

But now, the build system is using the committed iopc and always enable
this option, so it became useless. Remove it.

The iopc testsuite was not using this option, so a consequence is that
we have to update the expected typescript files (and we gain some code
coverage).

Change-Id: I9d1de7c7a339e507f07d7570a53eccac3d597e95
Priv-Id: a83e4e2d294ac0f28fa10b469cc3d8896d96f521