www: add type icQueries in iop.ts
A huge part of the mmsx codebase is dedicated to adding type to
the interfaces/rpc.
We can automate this task to make it more reliable and avoid time lost.

We first create a type for the icQuery function according
to every rpc declaration.
For example: the rcp 'insert' is called with the type 'insertIcQuery'.

We bundle every rpc's icQuery type in an interface in one 'IcQuery' type.
For example: the rcp 'insert' is inside the interface 'Query' so
'interfaces.Query.IcQuery' is a type that extends 'insertIcQuery'.

We rely on the tag 'backbone:iface' in the iop files to know
which iop object can use which interfaces.
For example: '@(backbone:iface, "bigdata.query.Query")'
placed before 'struct Report' indicate that the function icQuery inside
'Report_Model' extends 'interfaces.Query.IcQuery'.

When extending an object extends an other object,
it extends its icQuery type with its parent's one.
Since every parent must have an icQuery type that it may pass on to
a child object, we must define it with the smallest possible type.

If an iop object can't call any interface, we could use the type 'never'
to signify this impossibility but, this leads to some complex definition
in case of an inheritance.
We used '(args: never) => never' instead, as extending this type
is not an issue yet it still can alert a developer trying to use it.

Finally, some new interface must be imported from other files,
so we had to update the import declaration as well.
The path declared in the tag 'backbone:iface' being a char, we added
helpers on top of the file to parse it into an 'iopc_path'.
We also updated 'iopc_dump_import' to use it with theses new paths.

Change-Id: Ic076d0843fa25b07e5fa6780a1d272e1f12ec237
Priv-Id: 4320f1ca83c0508bf4e46f450878835e836fe467