eventloop: Fix tolerance compute issue.
The tolerance was computed based on repeat period, which means if we set
a timer to run in 600 ms each 30 min, the tolerance is the MIN of 3min
and 59s (the default timeout). So the first fire is in 1 minute instead
of 600 ms.

To fix this, we should compute the tolerance for the first fire with
`next`. And for the rest of the fires, we use the `repeat`.

Add a test for the first fire. For other cases, it seems difficult to
test automatically.

Change-Id: I8ee34c3c3884658e7ca286bc1bd22bc4a5f976a8
Priv-Id: eb4f01712ac82e1918788f9f80aa09b5e05de3b0