core: add max file age option to log file configuration
This commit add a new option to specify a max file age for log file.
We check the age of log files based on their creation date, and files
older than this limit are deleted. As we are using the file creation
date, we automatically add a margin corresponding to the value of
maxTime.

Change-Id: I6759532ef6540823dcd2671d36cde72f94d1d409
Priv-Id: 2490f62a11a2dd5c1fc73c72da0dfae467a9362a