qps: invalid assert on memory if there is no block
While performing fuzzing operations on QPS code, it has been detected
the following assertion:

    Assertion `qps->maps.tab[((__pg) >> 16)]' failed

Use case triggered here by fuzzing is the following one:
 - While calling qps_pg_remap_int() from qps_realloc_int(), a free block
   is available and fit the exact size expected (33 consecutive blocks)
   at the end of the map, starting at block 65503
 - The metadata of the next block is updated as a non free block (this
   metadata is at page 0 of the next map, which is valid even if there
   is no real map behind it: block 65503 + 33)
 - memory status of the block 65536 through PG_UNDEF macro is performed,
   which is referring to the beginning of the next map.

The analysis showed that the macro PG_UNDEF from qps_pg_remap_int()
function is actually called with a size of 0 block for this fuzzing
specific case, which seems not good. This could be without any effect if
the reference to the next map would have been valid, which is not the
case either. As a result, the assert "assert(map)" fails.

Prevent this case by ensuring memory property on some blocks can really
be set/changed.

Change-Id: If7c97acb5bab5f5085c5e9546918cbc755b030d4
Priv-Id: 414af598fd88844b42ca45d4f7665deb8ba8e0a5