clang: add robustness to clang cpp lib path
Depending on your Linux distribution the clang cpp library can be named
differently like:
- libclang-cpp.so.18.1
- libclang-cpp.so.14
- libclang-cpp.so -> libclang-cpp.so.18.1
The last symlink doesn't necessarily exist so it is better to be
explicit about the lib file we want to use.
If for example we are using LLVM 18.1.6
This patch would explicitely declare libclang-cpp.so.18 and if it cannot
be found in clang path it would fallback to libclang-cpp.so.18.1.
If neither is found, a fatal error is triggered.

Change-Id: I1426fdd8027e8df49e5432e2971057da95f9ffcf
Priv-Id: 6a44d95c9d43412f6008ce903b57809a90a09b34