net/http: change httpc_http2_ctx_t owner
Replaced the owner of the the httpc_http2_ctx_t object from http2_conn_t
to http2_client_t. Note the latter is an outer type of the former
effectively the ownership does not change. However, this choice make
more sense and parallels what is in the counterpart object
httpd_http2_ctx_t.

Change-Id: Icb720041deed1dbcb0e01bbfae2d49c592b4e7f7
Priv-Id: de18aa85ff919d4ebf3818d86ddcbff07e4cf051