net/tls: fix logger stopped when we got a SSL error.
net/tls: fix indents.

Change-Id: Ic3e76068231842b799700249f721f660da796287
Priv-Id: c39c6b7ea3c56643d3476651bc663db13da7575d

---

net/tls: update SSL error message.

Add the value of the SSL error in the log.

Change-Id: I90e6f6a2bef734e8a54908049f30950464ab2fe5
Priv-Id: 2119c5a29287734df9fdbf9b0d7855940525e3d6

---

net/tls: fix logger stopped when we got a SSL error.

When we got a SSL error, the logger was stopped.
The issue was that the SSL error queue was not cleaned.
Clean it after each error handling resolve the issue.

Change-Id: I060563caccdbf74d17c2935f45682ba5fabeaa59
Priv-Id: baf8e99f62e1bfeb994de270225519fb151a217b

---