qpsstress: implement fuzzing operations on QPS
* What
Based on the replay mechanism which can play QPS or Qhat/Qbitmap
operations, execute steps from a sequence generated by fuzzing engine.

* Why
This is hard to resolve potential issues around QPS and Qhat/Qbitmap
operations in a way that we can be more confident on the QPS source
code. In fact, we want to find tricky cases where QPS assert and/or QPS
corruptions cannot be easily spotted by a developer due to the
complexity of the code and the combination of operations.

Fuzzing can actually achieve that by generating entropy on operations
and exploring/monitoring all conditional branches based on them.

It actually confirmed unexpected behaviors/issues like seen in issues

* How
The following items are developed in this patch:
 - A mechanism based on the replay mechanism which takes a fuzzing
   binary blob as a sequence to play.
 - All previous random accesses are now predictable, as we need to
   replay fuzzing sequences without entropy (this is done by fuzzing).
 - Use define __fuzzing__ to distinguish fuzzing compilation from a
   "normal" one.
 - Ensure that no log are written during fuzzing operations except if
   debug is performed or crash file is replayed.
 - Ensure pure rubbish sequences are discarded to improve efficiency.

In addition to these items, we want to generate corpus files and a
dictionary containing fuzzing operations on request, so we don't waste
our CPU time for basic operations not known by fuzzing engine at
startup. The file "create-qpsstress-corpus.py" has been created to
create the initial fuzzing sequences.

Note that QPS snapshots and reopen are deactivated for now, as these
operations kill the whole purpose of fuzzing and the number of
executions per second (which is expected to be around 1000/sec for
efficient fuzzing).

Change-Id: Id4cbad220df8e905bacc270da54bd25b0ffdf407
Priv-Id: 82063e70dfc4c4e6e7c3426a3812252f8b78a004