http: do not send Content-Length with code 1xx or 204.
rfc7230 is clear :
A server MUST NOT send a Content-Length header field in any response
with a status code of 1xx (Informational) or 204 (No Content).

Most http1 and http2 clients are probably not strict and handles absence
or presence of Content-Length header in 204 (No Content) responses, but
this is not the case of Ericsson AMFs at Bytel.

rfc 7230 also says this :
 1.  Any response to a HEAD request and any response with a 1xx
       (Informational), 204 (No Content), or 304 (Not Modified) status
       code is always terminated by the first empty line after the
       header fields, regardless of the header fields present in the
       message, and thus cannot contain a message body.

which probably explains why most clients are tolerant, but nevertheless
our server must follow the spec.

Change-Id: I028bf3784ebd7d426ff79e394730d9d2e6909c4e
Priv-Id: a622eafb2f701ac983229dc8a190104ac76d18ea