build: enable march=x86-64-v2 compilation flag
Fix patch enables the -march=x86-64-v2 flag so we can benefit from more
“modern” hardware optimizations. The x86-64-v2 profile actually matches
the architecture level of the 2008 Intel Nehalem.

This is also the choice of Redhat for REHL 9, see:
https://developers.redhat.com/blog/2021/01/05/building-red-hat-enterprise-linux-9-for-the-x86-64-v2-microarchitecture-level

I think we could even move toward the x86-64-v3 architecture, which
match the 2013 Intel Haswell architecture, but it requires some
changes in the lib-common, unlike the v2 architecture, so we'll see
later.

See https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels for
more details.

Change-Id: Ic0f192ae069e839e5078e6a2cc3c46c7a6e75096
Priv-Id: 980c25d9d1a6c3aeb429faa3cfa5dc53b7cb90f5