waf/poetry: remove old packages from Poetry env
When updating the Poetry venv, we did not remove the old packages that
are no longer installed in the new version of the venv.
So when switching from 2022 to 2024, we ended up with an old version of
Poetry libs staying in Poetry venv (because we installed Poetry in
Poetry in 2022... and do not do that anymore in 2024, so the package has
not been updated).

Change-Id: I8cbddb7b28833f3dd4d8d33aa6e284c95195e011
Priv-Id: d10f85efb8cf07433b4ad959d20c510a0dbe8716