waf/poetry: do not install poetry when Python system
If asdf is configured to use the Python system, do not try to install
poetry neither.
Installing Poetry while using the system Python will add poetry libs in
the `~/.local/lib` directory of the user. The problem is that in 2024 we
don't manage poetry install anymore (poetry is managed by asdf), so we
end up with conflicting versions of Poetry (an old one in the libs of
the user, and a more recent one managed by asdf).
This problem arises on Debian 11 release builds only probably, because
Debian 11 is the only OS using the same Python version (Python 3.9) on
both 2022 and 2024.

Change-Id: I4ca0a7965b67fe0d4df2aa122d66dfe9246819c8
Priv-Id: de61db30ec283f9a53851ca2d4f2b534c61ae1b4