iop: move IOP environment from priv.h to iop.h.
In the next patches, the IOP environment will not longer be global and
will be freely manipulated.

So it will need to be declared in the public header iop.h and not in the
private one priv.h.

Document the different fields and functions of the IOP environment and
IOP object.

Change-Id: Ib0ba027643e858ef85e677682b8e1c7718af1659
Refs: #82540
Priv-Id: 162864f14cd2d2a9be7030b51dc918509d0e57d2