iop-dso: fix iop/dso for iop_env_t support.
The IOP DSOs are not heavily dependent of the IOP environment.

The lmid namespace is not stored in the IOP environment, and no longer
along with the IOP DSO.

This means that it is much easier to cleanly use a lmid namespace for
DSO for a particular IOP environment (we don't even need to unregister
the IOP environment and DSO when switching between DSOs).

Unfortunately, there is a drawback.
The glibc implementation supports a maximum of 16 namespaces per
process. After that limit, and error is thrown when trying to call
dlmopen():
    libc.so.6: cannot allocate memory in static TLS block

To mitigate this issue, use a pool of already opened namespaces that can
be reused.
This does not completely solve the issue if we open more than 15 DSOs
simultaneously.

Change-Id: I36fb49f1c1cebea379b6d0a31c5ef5cafc74990a
Refs: #82540
Priv-Id: e7c440b9e06a2d5ecb24fde75a326c2d2d0aa36c