tests/zparser: adapt parser with new screenshot function
We have changed the way screenshot are generate on Selenium test
But with this new behavior, the actual parser script cannot catch the
screenshot URL because screenshot does not appears anymore in traces.

Add a way to catch the screenshot on each scenario when available and
add it to global report. There is no possibility to have more than one
screenshot on each scenario.

Also add the fail step name to global report

Change-Id: I94a282bf0e03588829e7acf4a487e47f8092ddc4
Priv-Id: 77ef6dcfcce0c2a78af1114e1c15fdd8ca95dd69