core/str: add '*lstr_dupz()'
When dealing with files, we usually need null-terminated strings for
filenames, but in our code, we usually use 'lstr_t' for any string.

As a consequence, we often need to get a null-terminated string from an
lstr_t, and we cannot simply do use the "s" attribute because it is not
guaranteed to be null-terminated, even if it almost always is.

This patch adds the functions 't_lstr_dupz()' specifically for this kind
of use-case. It also adds 'lstr_dupz()' and 'mp_lstr_dupz()'.

Change-Id: I802d5380e56e10c4e5c2bb934071acc7667138ed
Priv-Id: 40e81f67deabf54850aa3d2651619215934f94d9