iopy: rework Iopy_make_plugin_from_handle into Iopy_make_plugin_iop_env.
The C exported function Iopy_make_plugin_from_handle() was never
actually used.
So we can freely rework it.

Convert it into Iopy_make_plugin_iop_env() to make it possible to create
a plugin from an existing IOP env.

The DSO can now be optional, and the modules are now loaded/unloaded
through the packages and not from the DSO anymore.

The IOP environment of the plugin is now initialized on __init__(), and
not in __cinit__() (called by __new__()), since the IOP environment can
be retained from an external source by Iopy_make_plugin_iop_env().

Ref: #82540
Change-Id: Ic7f1e8b3ead71b4c8cf6a0d2de72d75455430a35
Priv-Id: a5b4130871cc686c497d37a047f1396c3726ecf6