mem: Report leaked memory pools at shutdown.
We had a memory leak in all our mmsx components using MDE (cf. #104606),
because MDE leaked a whole (unused) mem ring pool at each computation.

And we failed to see this bug for years because nothing was reported by
asan. The reason is that because of the global dlists of memory pools
(used to print the memory pools state on kill -PWR), the pools and their
contents were considered as still reachable at shutdown.

So to fix that:

* remove all the mem pools from the global list in the shutdown
  functions of the associated modules => this is enough to reveal the
  leak with asan.

* also add a warning log at shutdown, if some memory pools are still
  alive, to ease leak detection when not using asan; exclude from this
  reporting global pools that we know are destroyed later.

Also rename the 'pool_list' fields in memory pools as 'pool_link' in all
memory pool types. This naming avoids the confusion between dlist lists
and links.

Change-Id: I0544ece520ba1ddd7c41c985ff0dd86ad007c867
Priv-Id: 149f68039871bf0000dbc1b64d2e64fd2f2fcf21