iopc: move internal stuff from iopc.h to iopc-internal.h
The iopc.h header is used in mmsx as the public header of iopc to use
the IOPC-DSO functions. But the iopc.h contains a lot of internal
definitions and functions that should absolutely not be exposed outside
of iopc.

So let's move all the non-public stuff to a new iopc-internal.h header.

Change-Id: Iecf467e3aa7ad1524eccecc0e9df8c67617991c2
Priv-Id: cf5fe1ac6299fa087b5e3c927b5ade4291213704