log: Preserve errno in logger_vlog.
I have observed that sometimes, the errno value displayed in the .debug
files generated by logger_panic is "success" in cases it is supposed to
be something else.

After investigation, I have understood that it happens because
logger_vlog does not preserve errno (that can be modified by the log
handler, I guess).

So use PROTECT_ERRNO in logger_vlog, which solves the bug.

Change-Id: I8107925589498272168cd4167733bd0d68bbd59a
Priv-Id: dbd4877a47f4064461f0624cd120a67d7a9a0d0e