error/log: Dump critical log in .debug files.
When a .debug file is generated by a call to e_panic/logger_panic (or by
e_fatal/logger_fatal in some rare cases), a useful information was
missing in the .debug file: the contents of this log.

The purpose of this patch is to add the contents of the critical log in
the generated .debug file, just like we print the current value of
errno.

To do that, e_panic & co write the log contents in a global variable
syslog_critical_log_g, that is then used in ps_write_backtrace.

Change-Id: I57008e1dfc080a370a96b5dad7f4f658c8adb196
Priv-Id: 2529db41ab99fc460054303f123a49871c6ec38a