iop-rpc-channel: reorder attributes of ichannel_t.
Some attributes of ichannel_t can or must be configured by the user,
whileothers are part of the life-cycle of the ichannel.

Before, there were no clear distinction between the two types of
attributes apart from the documentation.

Reorder the attributes and add a proper fold to split the two types of
attributes so it is easier for the user to use ichannel_t.

Change-Id: Ia36327e0867e4f9871a83d30f3c886897ae8d2e2
Priv-Id: 9d63c174a1a3107bcf306241bea5fb791c7e8f7b