iopy: fix package load dependency with modules.
When loading the packages, the order where the structs and the modules
are loaded actually matters.

When loading the modules, the python classes for the structs
corresponding to the arguments are created, but they are not inserted in
the python packages (this is the expected behaviour).

However, the interfaces can directly reference some structs that need to
be inserted in their python packages.

If we don't load the packages in the correct order, these structs will
not be correctly inserted in their python packages since they will be
considered as already created.

Before Ic7f1e8b3ead71b4c8c, for a given DSO, the modules were loaded
after loading every structs from all packages, so this was not an issue.

With Ic7f1e8b3ead71b4c8c however, the modules are loaded when loading
the packages. So, this triggered the issue.

We need to restore that order, and do the same when loading an IOP env
than when loading a DSO.

This also means that we cannot load one package at a time, but we need
to load a whole IOP env or IOP DSO.

So we need to convert Iopy_{add,remove}_iop_package() into
Iopy_{add,remove}_iop_dso().

Also solve some python exception handling.

Change-Id: I998f2dc0b32109a8a828e6eaf3c170d9c4e7c0e6
Priv-Id: f78b4820c9f9d0020bdbc591cbdcb65d2d966ec7