iop: forbid duplicate registration of IOP symbols.
Before, when using DSOs, we accepted registering two IOPs symbols with
the same type and name.

This was done like that because in older versions (before 2017), the
external DSOs had duplicate symbols.
This is no longer the case in our products, and the only case where it
happened was in the tests.

So we can safely remove this case and return an error instead.

Ref: #82540
Change-Id: I63fd3114d30a9d99935347215a19eab109387958
Priv-Id: 3b1215048021821a22c29187bdd6fd337b55a22a