core/mem: add a flag to disable leak detection for a mem pool
Pools that are allocated and freed using thread hooks are generally
freed *after* the part of the code that looks for leaked pool is run, so
the pool is mistakenly considered as leaked.

This patch adds a memory pool flag for these memory pools that we do not
want to check. It removes and replaces the previous named-based system:
we used to skip the checks for a given list of memory pool names.

Change-Id: I90e149389ce23c3d8d946d148d2be4759abb8ed7
Priv-Id: 848b7a5a6f3f83dc059ac97af99e81802579a83d