datetime: add methods to get complete date time and easy to read.
There is already some methods to get date with format ISO8601.
The only inconvenient of this format is that it could be easier to read.
Add a method letting us get a date with a format almost identical than
ISO8601 one but easier to read.
Date with format ISO8601: 2024-12-13T16:24:25Z
Date new method         : 2024-12-13 17:24:25 +0100

Change-Id: Iab18c51316e78a428da4d87d158f8b59c8e3ef0a
Priv-Id: 2defcae49f70e18e3a1249dc000b62f525a7a127