bit-wah: Remove dangerous mp_wah_(init|new)_from_data APIs.
These APIs are usually used with t_pool(), to spare a wah_wip/wah_delete
to the caller.

The problem is that in some very specific cases, the underlying function
calls qv_extend on qvectors that were init with qv_init_static on the
input pstream data. And in that case, the vectors are re-allocated on
the libc, and thus must be wiped.

So these APIs were broken by design, and must be removed.
wah_init_from_data and wah_new_from_data will be used instead.

Change-Id: I78a232c7c5b1df5505e747e8308f06620f911c43
Priv-Id: ecd27603d587a9fdaaae2a69b4e062069bff8a89