datetime: improve time_fmt_*_iso8601 methods.
When we print a timestamp, we check that the string got has the size
expected. As, this size is always the same, check the exact size is
better that check that the size is not bigger than the one expected.
Moreover, e_panic seems to be too brutal here. Do a e_error and assert
instead.

Change-Id: I5c051958c3ac859df9067debad0299a81bfaf135
Priv-Id: d91207efc15f76cfd52aff2f697630eacdae0e72