iop/rpc-channel: fix 100% CPU during handshake of stale connections
When an IChannel connection was established but went unresponsive during
the handshake phase, it resulted to a 100% CPU right *after* the soft
activity timer would be fired.

Indeed the handshake phase after writing its greeting message set the
IChannel in POLLIN only mode to wait the peer response. But when the
soft activity timer was fired before the end of the exchange, it
enqueued a IC_SC_NOP message to force some activity. The call to
ic_nop() resulted in the activation of POLLINOUT and thus a 100% CPU
since the IChannel was not ready to write any message yet.

The fix is done in two ways:
  - stop activating POLLINOUT in ic_queue() if the IChannel is not yet
    “connected” (is_connected flag);
  - stop enqueuing NOP operation if the queue is not empty which is
    useless.

Change-Id: I4be8cd1ca5ad15158d10e7362d94b32b0607b119
Priv-Id: 42c8422ec489e75d406c6cc06b2ceb710e8615e4