core/mem: add a flag to disable mem pool tracking
Our memory pools are usually listed in per-type global dlists so that we
can check their status with kill PWR (for instance). This feature is
meant to track the global pools (thread-local or not), but is not very
suitable for short-lived memory pools. This patch adds a flag that
disables this tracking.

Change-Id: I8e296f14c4e5153695bd4721b7328a7d546f95a1
Priv-Id: 69291231cb5c18b40109ec0991d82cd59534b8ad