net/hpack: enable hpack_decoder_extract to notify missing entries
A dynamic table (dtbl) miss is when we receive an index that we do not
have in our dynamic table.
In a nominal setup, such as an HTTP/2 server, this should not be an
issue. But in the probe use case this is something that we have to deal
with.

This commit allows hpack_decoder_extract to return
HPACK_PARSER_STATUS_MISSING_ENTRY when the dtbl idx is missing.
This allows for the caller to handle this case.

For http.c it does not change anything as it is called with RETHROW,
which will trigger in case of ERROR or MISSING_ENTRY.

Furthermore, the XHDR_* enums are now public and renamed HPACK_XHDR_*
to facilitate the caller's handling of missing entries.

Change-Id: I571fcbddbcd42235ca7dfbd137ed5a49104429b7
Priv-Id: 3cca704981931a7f24a05915d5510ff29f85c176