core: introduce helpers dedicated to compiler warnings
The patch introduces a few helpers to ease and standardize the way we
deal with compiler warnings. Indeed we have a lot of different usages,
and it's always a pain to check which GCC version can use diagnostic
push/pop or if a warning is known.

For more details, look at the added documentation in core-macros.h.

Change-Id: I34087c15604183f9f268e75136fa8cb4b042e477
Priv-Id: f832d4efc241faef60f962b778713e5aaefdf951