thr: Add thr_cfg_t, to configure max parallelism.
This structure is not the optional argument of the thr module. When set,
it allows to override the max parallelism instead of taking the number
of CPUs. The THR_MAX_PARALLELISM environment variable still prevails.

Also, by default don't go higher than 64: we have a very low level of
confidence with that much threads.

Change-Id: Ib538a8e87693ba705512cd56ce1133f236e20c35
Priv-Id: 3b7def2e98e8824a4eb3a7ae4ab62df0d1a693b9