iop: split 'iop_env_t' and make it opaque.
In 'iop_env_transfer()' we make some swap between two IOP envs.
The issue is that we need to be careful about which fields need to be
swapped, and thus rely on comments to avoid the mistakes.

This was not really robust.

Instead, we can extract the swapped fields of 'iop_env_t', and put them
in a sub-structure 'iop_env_ctx_t'.
This structure is swapped completely in 'iop_env_transfer()'.

'iop_env_t' is now an opaque read-only structure, except from the
available setters.

This has the benefit of possibly adding some locking mechanism in the
future to make 'iop_env_t' thread-safe.

Also, remove exporting `iop_dso_load_handle()` and put it as static in
dso.c since it is no longer used by IOPy with a5b4130871cc686c.

Change-Id: Iae940895703ca12a0981f55261bf348de6f765b5
Priv-Id: b2e2acb396c4ad393a8d0f8e55be371c08e52f29