core/str-buf: assert on overlaps for sb_add/sb_set
When using sb_set/sb_add, we do not check if the buffer is overlapping
with the sb even though we always use memcpy on that buffer.
This can lead to bugs when the user is not being careful about overlaps.

This commit adds asserts in sb_add and sb_set to verify that the user is
not manipulating overlapping buffers.
The asserts are accompanied with messages to explain to the user the
proper way to handle their overlapping buffer:
- in case of sb_set, an overlapping buffer should be contained whole in
  the sb, so using sb_shrink/sb_skip and/or sb_clip is the proper way to
  handle this.
- in case of sb_add, an overlapping buffer should be allocated elsewhere
  (where it does not overlap with the sb) before using sb_add.

Change-Id: I40d8cefc057fadfb2e4ae0fead872ee1270b0c51
Priv-Id: 29b39d7244d4317acf6cc4c04293c4ed64684afd