httpc: get rid of on_busy callback
The on_busy callback of the HTTP client library was never used but made
the code really confusing.

In `httpc_pool_get`, when no connection is available, a new connection
is created and NULL is returned, because the caller has to wait of the
actual connection establishment to use it (and retrieve it with a
subsequent `httpc_pool_get` call).

But because of the `on_busy` callback, the code handled the eventual
case of the connection being magically connected by the callback which
makes no sense.

So we just remove the callback and simplify the code.

Change-Id: I54f9d54591c26ea7fd7dba52b06f439100679eb1
Priv-Id: 03a78dd0770396abe357c80a7ed2ecf66af77122