iop/rpc-channel: fix ic_get_client_addr with unix socket
On the server side of a unix socket, getpeername provides an empty
su->sunix.sun_path and we end up with "unknown unix socket" returned by
t_addr_fmt_lstr in ic->client_addr

Then addr_info is used on ic->client_addr to validate the address which
is an invalid and useless DNS query that can be avoided.

Solution: always set client_addr to 127.0.0.1 for unix sockets (like in
ic_set_local), that way DNS resolution is valid.

Change-Id: I3497d9b9608877d73750a0e3cad3362d1de6f8c6
Priv-Id: 34a6fdbbdab45e33641d488f990adcf1a5ffc07b