rpc-http-client: http_iop_channel_cancel_messages.
When a channel is deleted, if there's a message waiting to be connected,
it's lost.

This is a problem because, in some cases, we need the message life cycle
to be complete.

To do this, we add a function that calls the message callback to cancel
it.

Messages could be cancelled when the channel is deleted, but the impact
would have to be checked.

Change-Id: Ifd8f56be767263fa1807a2d3bb7ff2705cbbf009
Priv-Id: a3cf108ac143addbe1191925cf1db7bfbef3fb55