net/http: support bearer auth in httpd_trigger_t
In order to realize the user story #109359 related to OAuth2
authentication, httpd server should now be able to check credentials
from both:
 - basic authentication,
 - bearer authentication.

To achieve that, the function httpd_check_credentials looks for the
"Authorization" field in HTTP header and defines which
authentication callback must be triggered, based on the fact that basic
auth and/or bearer auth callbacks are defined.

Note that empty credentials are allowed (i.e. no Authorization field in
http header), meaning that user, password or token can be set to a NULL
pstream when calling the check credential callback.

Change-Id: I004c250fa793377c005ca2b1170514c63340e5e8
Priv-Id: ce3e132ae45a2457c550371585a82091de08f846