iop: add iop_struct_get_field() function.
Find an IOP field description for a IOP struct definition and field path.

It is similar to iop_field_find_by_name(), but takes a full IOP path
instead of just the field name.

The implementation is factorized with iop_get_field_const() but it does
not take an object instance and only check the object description.

Also fix iop_get_field() and iop_get_field_const() non-null ptr. An
error is directly thrown when `ptr` was null, so `ptr` is not expected
to be null.

Change-Id: I493dfc92ddf9cc3c60033ad0023f7889042f2f55
Priv-Id: 6c2d130df7f11dbded6cd4af47ed03c3ce6d8fe4