iopc: allow '@private' to apply on void required fields.
In order to mask a field from documentation, '@private' can be applied
onto the field.
Since the field is not documented, the user should not have to explicit
set it (they will not know the field at all), and thus the field cannot
be required.

However, required void fields are a special case that do not require
some value.
So we should be able to apply '@private' on required void fields.

Also refactor and factorize some functions in iopc-parser.c.

Change-Id: I2af76bed4a6efb9d116835a877d3b346c9b44d32
Refs: #8536
Priv-Id: 1d0e414b8c0d48a76ae3f84354daa6360f8c667a