z: Also filter blk tests according to z_register_exports prefixes.
Standard tests (exported with Z_GROUP_EXPORT) are registered with
z_register_exports according to a prefix (that is generally the
directory where the zchk program belongs). This way, only the tests are
are actually in the zchk directory are executed, and not the tests that
are defined somewhere else (which can happen if tests are present in
used libraries).

The problem is the block test groups (exported with Z_BLKGROUP_EXPORT)
are registered when exported, and thus are not filtered-out by the
prefix.

This patch fixes this issue, by storing the prefixes in a global vector
in z_register_exports, and then filter-out groups upon execution based
on the registered prefixes.

Change-Id: Iac94b52b780c025aec9da5ea839e9b88d83a6cc3
Priv-Id: 966ce7c0bfd7b9ad8aa9db70ce79437e982d8ff8