waf[rust]: Fix check_rust, reimplement check_rust_source.
check_rust was trying to set the rust tgen features from a
before_method('process_rule'), which is apparently not possible.

So have a build pre_fun instead, that runs through all the rust tgens
and sets their features, before they are posted. Also move the tgen
checks there.

Now, the 'rust' feature seems unused to waf, so we have to define a
dummy method using it to avoid a warning from waf.

Change-Id: Id91adfc0ba25ab43bed6377b1563233db94c969d
Priv-Id: 0a1bc1b179de6d4abd0340bb598e3b2e265216bf