waf: In compile_fpic, ensure created tgens are posted in the good group.
Because compile_fpic is executed after the build() method, it is
executed in the context of the code_compiling group. As a consequence,
all the tgens in generates belong to the code_compiling group, which is
the last executed one.

This is a problem if some shared libraries are defined in other groups,
because they would need fpic libraries that are built later (too late).

This patch ensures that the generated tgens belong to the same group as
the tgens that need them.

Change-Id: Ie635620935a112352516a36f87730c5f915264ad
Priv-Id: 19d369495a0b38835685524e01bea2e412145925