mypy/iopy: fix basic types deductions for base IOPy types.
Fix and add `__getattr__()` methods to base IOPy classes.
This way, we will not have the different 'attr-defined' when using
non-stub typing with base classes (e.g. iopy.Package).
This does not make the code really more safer when using base classes,
but this is easier to use.

* Every unknown attribute of a Plugin is a Package.
* Every unknown attribute of a Package is a EnumBase or StructUnionBase.
* Every unknown attribute of a StructUnionBase is Any.
* Every unknown attribute of a Channel is an Interface.
* Every unknown attribute of a Interface is a RPC.

For StructUnionBase, allow any attributes to be set with `__setattr__()`
because they can potentially be some IOP fields.

For stub typed classes, `__getattr__()` (and `__setattr__()` for
StructUnionBase) is reset to None to keep the
correct strict detection of valid fields.

Try to auto-deduce the type of an object created from a `Basic` class.
`Basic` cannot be instantiated on run-time, by the type checker can
still use it, and `Package.__getattr__()` returns a `type[Basic]`.
Depending on the arguments passed onto the constructor, we can try to
deduce the type of the created object:
- If we try to do a copy of the IOPy object, return the same type of the
  object (EnumBase, StructBase or UnionBase).
- If there is only one argument of type `str` or `int`, we assume we try
  to create an enum (EnumBase). Technically, we can also create a union
  with an unambiguous type, but it is less likely.
- If there is only one argument of type dict, we try to either create a
  union or a struct (StructUnionBase).
- If we use keyword arguments, we try to either create a union or a
  struct (StructUnionBase).
- Otherwise if there is only argument, we try to create a union with an
  unambiguous type (UnionBase).

Add some tests to test this trick (tested statically with mypy).

Change-Id: I0a433c4831202739d4f8b08f50c039c576d0ddac
Priv-Id: d76b6c7ecde35efa36a94353b9701d4bc73a3465