waf: Add a mechanism to retry failed tests.
When performing the initial 'waf check' or 'waf fast-selenium' on
buildbot, any failure on these checks will directly invalidate the
build and a new build must be triggered to check that failures are not
random ones.

The purpose of this patch, associated with another buildbot patch, is
to run dedicated tests to retry. It is divided in 3 parts:

 1. A switch in file retry_checks.py has been introduced, based on waf
    commands named 'check-retry' or 'fast-selenium-retry'. If one of
    these commands is seen, we just switch to the new script described
    below.
 2. A file named retry_checks.py has been introduced, and behaves
    the same way as list_checks.py does: it generates an output to
    stdout, so run_checks.sh can run tests based on each line printed.
 3. New waf "rules" in build/waftools/common.py have been added

Note that this retry mechanism can be used by developers too, assuming
that they are at the mmsx root folder, like the following example:

    Z_LIST_RETRY=one/ci/features/multi-db.feature,one/zchk-message.py
    waf check-retry

Change-Id: I48b0a1122f045e01b405d582138b55e4fb187952
Priv-Id: 4db56174d88569746e4ec3e2dfffca905b62c230