rust: rework waf rust cargo package tasks.
Cargo works by package.
One package can only contain one lib and/or multiple bins.
The build.rs of a package is only run once for the lib and/or multiple
bins of a package.

With these constraints, this is not possible to have one waf target per
rust bin, and multiple waf targets for the same cargo package.
We cannot know which rust bin we try to build when building the cargo
package in the build.rs.

Instead, we need to have another approach.
We use one waf target per cargo package.

- All the targets of a cargo package are build when building the waf
  target.
- All the bins and shlibs of the cargo package are hard-link in the
  directory of the waf target.
- The waf target name must refer to the cargo package name.
- One one '_waf_build_env.json' is used by cargo package and waf target.

Change-Id: If317790522f1d3b9c96e78efa65c6e9547fcf878
Priv-Id: fa62918da7338b487bccbc636b01e877cfb75bb3