build: Fix dirty detection in version.sh.
To know if there was some diff in the committed files (to set the
"dirty" flag or not), version.sh was using "git diff-files --quiet".

The problem is that diff-files seems to base its detections on the
modification date of the files and not on their actual content. So for
example, if a file is re-written identically (which happens in mmsx with
"waf www" on the po files), they were considered as modified by
diff-files and the "dirty" flag was wrongly set.
Because of that, we had constant re-links during the "waf check" step on
some buildbots (probably depending on the git revision).

This commit replaces "diff-files" by "diff" which does not have this
optimization. This is probably a bit slower, but at least it is correct
and in the end it will save a lot of compilation time.

Change-Id: I1b441e816cd4e63fb3b8c0b256dae5ff0667d215
Priv-Id: 1b2a23d1249d85ebdbf74420bcc537d3baf3baf5